requirements.txt
torch
transformers
huggingface_hub
nltk
datasets
openai-whisper
torch
assemblyai
ollama
elevenlabs
mpv
portaudio